:bug: Do not call Mongo if no workspaceId is provided on deleteUnlinkedPages.

+++ services/workspaces/src/services/workspaces/crud/pages.ts
@@ -280,6 +280,9 @@ class Pages {
   };
 
   deleteUnlinkedPages = async (workspaceId: string, knownIds: string[]) => {
+    if (!workspaceId) {
+      return;
+    }
     const Pages = await this.accessManager.model(SubjectType.Page);
     try {
       await Pages.deleteMany({
